+++ R/RcppExports.R
@@ -2007,15 +2007,15 @@ hydraulics_E2psiTwoElements <- function(E, psiSoil, krhizomax, kxylemmax, n, alp
 #' @param psiLeaf Leaf water potential (in MPa).
 #' @param c,d Parameters of the Weibull function.
 #' @param P50,slope Parameters of the Sigmoid function.
-#' @param PLC_crit Critical leaf PLC corresponding to defoliation
-#' @param P50_cv Coefficient of variation (in percent) of leaf P50, to describe the
+#' @param criticalLeafPLC Critical leaf PLC corresponding to defoliation
+#' @param cvLeafP50 Coefficient of variation (in percent) of leaf P50, to describe the
 #' variability in hydraulic vulnerability across crown leaves.
 #' 
 #' @details The functions assume that crowns are made of a population of leaves whose
 #' hydraulic vulnerability (i.e. the water potential corresponding to 50% loss of conductance) 
-#' follows a Gaussian distribution centered on the input P50 and with a known coefficient of variation (\code{P50_cv}).
+#' follows a Gaussian distribution centered on the input P50 and with a known coefficient of variation (\code{cvLeafP50}).
 #' The slope parameter (or the c exponent in the case of a Weibull function) is considered constant.
-#' Leaves are hydraulically disconnected, and shedded, when their embolism rate exceeds a critical value (\code{PLC_crit}).
+#' Leaves are hydraulically disconnected, and shedded, when their embolism rate exceeds a critical value (\code{criticalLeafPLC}).
 #' 
 #' @return The proportion of crown defoliation.
 #' 
@@ -2029,14 +2029,14 @@ hydraulics_E2psiTwoElements <- function(E, psiSoil, krhizomax, kxylemmax, n, alp
 #' 
 #' @name hydraulics_defoliation
 #' @keywords internal
-hydraulics_proportionDefoliationSigmoid <- function(psiLeaf, P50, slope, PLC_crit = 0.88, P50_cv = 10.0) {
-    .Call(`_medfate_proportionDefoliationSigmoid_c`, psiLeaf, P50, slope, PLC_crit, P50_cv)
+hydraulics_proportionDefoliationSigmoid <- function(psiLeaf, P50, slope, criticalLeafPLC = 0.88, cvLeafP50 = 10.0) {
+    .Call(`_medfate_proportionDefoliationSigmoid_c`, psiLeaf, P50, slope, criticalLeafPLC, cvLeafP50)
 }
 
 #' @name hydraulics_defoliation
 #' @keywords internal
-hydraulics_proportionDefoliationWeibull <- function(psiLeaf, c, d, PLC_crit = 0.88, P50_cv = 10.0) {
-    .Call(`_medfate_proportionDefoliationWeibull_c`, psiLeaf, c, d, PLC_crit, P50_cv)
+hydraulics_proportionDefoliationWeibull <- function(psiLeaf, c, d, criticalLeafPLC = 0.88, cvLeafP50 = 10.0) {
+    .Call(`_medfate_proportionDefoliationWeibull_c`, psiLeaf, c, d, criticalLeafPLC, cvLeafP50)
 }
 
 #' @rdname hydraulics_scalingconductance