+++ src/hydraulics_c.cpp
@@ -449,15 +449,15 @@ double E2psiTwoElements_c(double E, double psiSoil, double krhizomax, double kxy
 //' @param psiLeaf Leaf water potential (in MPa).
 //' @param c,d Parameters of the Weibull function.
 //' @param P50,slope Parameters of the Sigmoid function.
-//' @param PLC_crit Critical leaf PLC corresponding to defoliation
-//' @param P50_cv Coefficient of variation (in percent) of leaf P50, to describe the
+//' @param criticalLeafPLC Critical leaf PLC corresponding to defoliation
+//' @param cvLeafP50 Coefficient of variation (in percent) of leaf P50, to describe the
 //' variability in hydraulic vulnerability across crown leaves.
 //' 
 //' @details The functions assume that crowns are made of a population of leaves whose
 //' hydraulic vulnerability (i.e. the water potential corresponding to 50% loss of conductance) 
-//' follows a Gaussian distribution centered on the input P50 and with a known coefficient of variation (\code{P50_cv}).
+//' follows a Gaussian distribution centered on the input P50 and with a known coefficient of variation (\code{cvLeafP50}).
 //' The slope parameter (or the c exponent in the case of a Weibull function) is considered constant.
-//' Leaves are hydraulically disconnected, and shedded, when their embolism rate exceeds a critical value (\code{PLC_crit}).
+//' Leaves are hydraulically disconnected, and shedded, when their embolism rate exceeds a critical value (\code{criticalLeafPLC}).
 //' 
 //' @return The proportion of crown defoliation.
 //' 
@@ -473,9 +473,9 @@ double E2psiTwoElements_c(double E, double psiSoil, double krhizomax, double kxy
 //' @keywords internal
 // [[Rcpp::export("hydraulics_proportionDefoliationSigmoid")]]
 double proportionDefoliationSigmoid_c(double psiLeaf, double P50, double slope, 
-                                     double PLC_crit = 0.88, double P50_cv = 10.0) {
-   double P50_crit = psiLeaf - (log((1.0 - PLC_crit)/PLC_crit)/(slope/25));
-   double pnorm =  normal_cdf(P50_crit, P50, std::abs((P50_cv/100.0)*P50));
+                                     double criticalLeafPLC = 0.88, double cvLeafP50 = 10.0) {
+   double P50_crit = psiLeaf - (log((1.0 - criticalLeafPLC)/criticalLeafPLC)/(slope/25));
+   double pnorm =  normal_cdf(P50_crit, P50, std::abs((cvLeafP50/100.0)*P50));
    double PDEF = (1.0-pnorm);
    return(PDEF);
 }
@@ -484,11 +484,11 @@ double proportionDefoliationSigmoid_c(double psiLeaf, double P50, double slope,
 //' @keywords internal
 // [[Rcpp::export("hydraulics_proportionDefoliationWeibull")]]
 double proportionDefoliationWeibull_c(double psiLeaf, double c, double d, 
-                                       double PLC_crit = 0.88, double P50_cv = 10.0) {
-   double d_crit = psiLeaf/pow(-1.0*log(1.0 - PLC_crit), 1.0/c);
+                                       double criticalLeafPLC = 0.88, double cvLeafP50 = 10.0) {
+   double d_crit = psiLeaf/pow(-1.0*log(1.0 - criticalLeafPLC), 1.0/c);
    double P50 = xylemPsi_c(0.5,1.0, c, d);
    double P50_crit = xylemPsi_c(0.5,1.0, c, d_crit);
-   double pnorm =  normal_cdf(P50_crit, P50, std::abs((P50_cv/100.0)*P50));
+   double pnorm =  normal_cdf(P50_crit, P50, std::abs((cvLeafP50/100.0)*P50));
    double PDEF = (1.0-pnorm);
    return(PDEF);
  }