+++ src/phenology_c.cpp
@@ -134,9 +134,9 @@ void updateLeaves_c(ModelInput& x, double wind, bool fromGrowthModel) {
           x.above.LAI_dead[j] += LAI_exp_prev;//Check increase dead leaf area if expanded leaf area has decreased
           x.internalPhenology.leafSenescence[j] = false;
         } else {
-          if(x.control.commonWB.cavitationInducedDefoliation) {
+          if(x.control.defoliation.cavitationInducedDefoliation) {
             double LAI_exp_prev= x.above.LAI_expanded[j]; //Store previous value
-            x.above.LAI_expanded[j] = x.above.LAI_live[j]*std::min(x.internalPhenology.phi[j], 1.0 - proportionDefoliationWeibull_c(psiLeafPLC, x.paramsTranspiration.VCleaf_c[j], x.paramsTranspiration.VCleaf_d[j], 0.88, 10)); //Update expanded leaf area (will decrease if LAI_live decreases)
+            x.above.LAI_expanded[j] = x.above.LAI_live[j]*std::min(x.internalPhenology.phi[j], 1.0 - proportionDefoliationWeibull_c(psiLeafPLC, x.paramsTranspiration.VCleaf_c[j], x.paramsTranspiration.VCleaf_d[j], x.control.defoliation.criticalLeafPLC, x.control.defoliation.cvLeafP50)); //Update expanded leaf area (will decrease if LAI_live decreases)
             x.above.LAI_dead[j] += std::max(0.0, LAI_exp_prev - x.above.LAI_expanded[j]); // Add senescence leaves to dead
           } else {
             x.above.LAI_expanded[j] = x.above.LAI_live[j]*x.internalPhenology.phi[j]; //Update expanded leaf area (will decrease if LAI_live decreases)
@@ -144,9 +144,9 @@ void updateLeaves_c(ModelInput& x, double wind, bool fromGrowthModel) {
         }
       } else {
         //Apply defoliation effects to evergreens
-        if(x.control.commonWB.cavitationInducedDefoliation) {
+        if(x.control.defoliation.cavitationInducedDefoliation) {
           double LAI_exp_prev= x.above.LAI_expanded[j]; //Store previous value
-          x.above.LAI_expanded[j] = x.above.LAI_live[j]*(1.0 - proportionDefoliationWeibull_c(psiLeafPLC, x.paramsTranspiration.VCleaf_c[j], x.paramsTranspiration.VCleaf_d[j], 0.88, 10));
+          x.above.LAI_expanded[j] = x.above.LAI_live[j]*(1.0 - proportionDefoliationWeibull_c(psiLeafPLC, x.paramsTranspiration.VCleaf_c[j], x.paramsTranspiration.VCleaf_d[j], x.control.defoliation.criticalLeafPLC, x.control.defoliation.cvLeafP50));
           x.above.LAI_dead[j] += std::max(0.0, LAI_exp_prev - x.above.LAI_expanded[j]); // Add senescence leaves to dead
         } else {
           x.above.LAI_expanded[j] = x.above.LAI_live[j];