+++ NEWS.md
@@ -1,6 +1,7 @@
 # medfate 5.1.0
 * Compilation patch for C++ 20
 * Mistletoe effects on water balance
+* New control parameters related to cavitation-induced crown defoliation
 * Recovery from defoliation depends on relative growth rate of buds 
 * Water potential at turgor loss point as estimation of Gsw_P50_Baldocchi
 * SpParamsMED now is filled with package traits4models